// src/app/core/models/tour.ts
export interface Tour {
  id: number;
  name: string;
  description: string;
  price: number;
  duration: string;
  image: string;
  startDate: string;
  endDate: string;
  // rating: number;
  // totalSeats: number;
  // availableSeats: number;
  location: string;
  status: string;
  // createdAt: string;
  // updatedAt: string;
  category: string;
  // originalPrice: number;
  // discountedPrice: number;
  discount: number

}